Start moving screenshots initiative: kde-color-scheme-screenshot
I had the idea to move the screenshots from the various blue-mood
repos to this one, in order to reduce the size of said repos for
anyone wishing to clone them (presumably, no-one will want to clone
this repo), inspired by
https://gitlab.com/protesilaos/tempus-themes-screenshots and
https://gitlab.com/protesilaos/dotfiles-screenshots. I'll first copy
each screenshot to the new location and change the README link
accordingly, and then will delete it from the original repo, linking to
the screenshot's new location instead. I'm starting off with kde-color-scheme.